+++ src/pages/LandingPage/LandingPage.style.ts
@@ -1,0 +1,38 @@
+import { css } from '@emotion/react';
+
+export const landingContainer = css({
+  width: '100vw',
+  display: 'flex',
+  flexDirection: 'column',
+  alignItems: 'center',
+});
+
+export const recommendPostWrapper = css({
+  width: '80%',
+  display: 'flex',
+  flexDirection: 'column',
+  alignItems: 'center',
+});
+
+export const morePostWrapper = css({
+  width: '100%',
+  display: 'flex',
+  flexDirection: 'column',
+  alignItems: 'center',
+  marginBottom: '20px',
+  marginTop: '50px',
+});
+
+export const headingWrapper = css({
+  display: 'flex',
+  justifyContent: 'flex-start',
+  width: '80%',
+  marginBottom: '20px',
+});
+
+export const headingWithButtonWrapper = css({
+  display: 'flex',
+  justifyContent: 'space-between',
+  alignItems: 'baseline',
+  width: '70%',
+});
